Move `mkPeers` (now: `choosePeers`)
This just moves a bit of code around; it doesn't actually take advantage
of the change yet.
